Added routing to existing containers
