+++ package.json
@@ -10,6 +10,8 @@
     "material-icons-react": "^1.0.4",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.0",
     "react-table": "^7.0.0",
     "react-table-6": "^6.11.0"
