+++ src/components/layout/layout.js
@@ -1,13 +1,10 @@
 import React from 'react';
-import Navbar from '../navbar/Navbar';
+
 
 const layout = (props) => (
-    <div>
-        <Navbar user="nombreUsuario" />
-        <main className="container">
-            {props.children}
-        </main>
-    </div>
+    <main className="container">
+        {props.children}
+    </main>
 );
 
 export default layout;
