--- src/containers/tables/administracion/Clientes.js
@@ -1,24 +1,0 @@
-import React, { Component } from 'react';
-import Title from '../../../components/ui/title/Title';
-import Table from '../Table'
-
-class Clientes extends Component {
-
-    render() {
-        return (
-            <div>
-                <Title
-                    title="Busqueda de Clientes" titleType="title-table" />
-                <Table
-                    url="https://jsonplaceholder.typicode.com/posts"
-                    header1="Código" accessor1="id" filterable1={true} width1={80}
-                    header2="Nombre" accessor2="title" filterable2={true} width2={300}
-                    header3="Estado" accessor3="userId" filterable3={false} width3={150}
-                    header4="Correo Electrónico" accessor4="body" filterable4={false} width4={400}
-                    btnTxt="Agregar Nuevo Cliente" />
-            </div>
-        );
-    }
-}
-
-export default Clientes;
